Update the history display.
In particular, give a green highlight to letters when they are correct.
This is reminiscent of both wordle and the matrix. And I think helps to
show the "story" of how it was solved.

Note that the highlights will be wrong if you have not yet solved the
cryptogram, or if you've typed in the cryptogram incorrectly.