Begin keeping a log of actions.
My first application of this is to easily know which cryptograms I've
previously worked on. I can use these for future tests.

Currently, the log may not always be sufficient to recover from an
interrupted session. In particular, a shuffle will lose progress. There
may be other cases I haven't thought of. In addition, I have not
implemented anything to actually resume a previous session.