Add the ability to show bigram freq tables.
I thought these might be helpful, but based on a few examples, I
actually think they're not useful until the cryptogram gets long enough.
Basically the tables tend to look disimilar for text as short as a
sentence or two. Still, I decided to leave in this feature. It was a fun
little exercise to derive some version of the bigram frequencies for
English. I used the text of David Copperfield to do so.